Address review feedback: split modules and minor fixes
Move RemoteAssignmentFetcher, RemoteAssignmentResult, and
get_remote_assignments back to resource_registry/tasks/sync.py —
remote-fetching is a syncing concern, not an RBAC concern.
assignment_utils.py now contains only local-model code.

Additional fixes from review:
- logger.warning → logger.error for missing org/team Resources
- Add ValueError guard in get_content_object for content_type=None
- Add tests for _paginate exception path and object_id fallback
